4.94

+++ src/main/java/jp/co/sony/ppog/entity/City.java
@@ -30,7 +30,7 @@ public class City implements Serializable {
 	 */
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE)
-	private Integer id;
+	private Long id;
 
 	/**
 	 * This field corresponds to the database column NAME
@@ -74,7 +74,7 @@ public class City implements Serializable {
 	 *
 	 * @return id
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 
@@ -83,7 +83,7 @@ public class City implements Serializable {
 	 *
 	 * @param id セットする id
 	 */
-	public void setId(final Integer id) {
+	public void setId(final Long id) {
 		this.id = id;
 	}
